project-05-financial-intelligence/pyproject.toml
[project]
name = "financial-intelligence"
version = "0.1.0"
description = "AutoGen SelectorGroupChat financial intelligence with regime detection and Phoenix observability"
requires-python = ">=3.11"
dependencies = [
    "autogen-agentchat>=0.4.0",
    "autogen-ext[openai]>=0.4.0",
    "yfinance>=0.2.48",
    "numpy>=1.26.0",
    "pandas>=2.2.0",
    "arize-phoenix>=7.0.0",
    "opentelemetry-instrumentation-openai>=0.40.0",
    "openai>=1.50.0",
    "httpx>=0.27.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "websockets>=12.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
